backend/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
dnspython==2.3.0
psycopg2==2.9.5
pymongo==4.3.3
pytz==2022.7.1
sqlparse==0.2.4